feat(api): add hscode api router using proxy
issue #36
